requirements.txt
flask==2.0.1
python-dotenv==0.19.0
openai==1.3.0
Pillow==9.5.0
werkzeug==2.0.1
pdf2image==1.16.3
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
gunicorn==21.2.0
